Use trait to refactor likable models.
