+++ app/Models/Comment.php
@@ -2,21 +2,12 @@
 
 namespace App\Models;
 
+use App\Traits\Likable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Comment extends Model
 {
     use HasFactory;
-
-    public function like($user = null)
-    {
-        $user = $user ?: auth()->user();
-        return $this->likes()->attach($user);
-    }
-
-    public function likes()
-    {
-        return $this->morphToMany(User::class, 'likable')->withTimestamps();
-    }
+    use Likable;
 }
