+++ app/Models/Post.php
@@ -4,10 +4,12 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use App\Traits\Likable;
 
 class Post extends Model
 {
     use HasFactory;
+    use Likable;
 
     public $fillable = [
         'user_id',
@@ -25,15 +27,6 @@ class Post extends Model
         return $this->belongsToMany(Tag::class)->withTimestamps();
     }
 
-    public function like($user = null)
-    {
-        $user = $user ?: auth()->user();
-        return $this->likes()->attach($user);
-    }
 
-    public function likes()
-    {
-        return $this->morphToMany(User::class, 'likable')->withTimestamps();
-    }
 
 }
